Delete remaining typed structs and greatly reduce the amount of state in the client.
Yes, this means the client no longer knows anything about guilds, members, channels, etc. by default.
If you need it, fetch it with the REST API. If you need caching, feel free to build a caching layer on
top of this client's gateway events.

This greatly reduces the amount of code in this library, so I don't have to touch it as often again.